refactor(export): isolate sticker tracking validation
